Share one conformance output matcher across both JS runners
stub-tool-axes.test.ts carried a stripped copy of matchesOutput that
silently dropped JsonContains handling and the Rust (?s) to JS dotall
translation — a case using either would diverge from the Rust harness.
Both suites now import the single full implementation from
matches-output.ts. Also fall back to copy+chmod on win32 when symlink
creation is not permitted.
